rename environments to environment, add to first-level import
